add sweetalert

+++ SON/admphp/templates/admin/pages/view.tpl.php
@@ -1,4 +1,4 @@
-<h3 class="mb-5">Administração de Páginas</h3>
+<h3 class="mb-5">Detalhes da Páginas</h3>
 
 <div class="row">
   <div class="col-3">
@@ -9,7 +9,7 @@
       <dt class="col-md-4">URL</dt>
       <dd class="col-md-8">/ - <a href="http://" target="_blank"></a>Abrir</dd>
 
-      <dt class="col-md-4">Criaco em </dt>
+      <dt class="col-md-4">Criado em </dt>
       <dd class="col-md-8">2020-08-23 16:00</dd>
 
       <dt class="col-md-4">Atualizado em </dt>
@@ -23,6 +23,6 @@
 </div>
 <p>
   <a href="/admin/pages/1/edit" class="btn btn-warning">Editar</a>
-  <a href="/admin/pages/1/delete" class="btn btn-danger">Remover</a>
+  <a href="/admin/pages/1/delete" class="btn btn-danger confirm">Remover</a>
 </p>
 <a href="/admin" class="btn btn-primary">Voltar</a>
